ai-control/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-evdev>=1.6.0
Pillow>=10.0.0
websockets>=12.0
pydantic>=2.0.0
tomli>=2.0.0;python_version<"3.11"
sdnotify>=0.3.0
aiohttp>=3.9.0
